update: combine mantine custom fonts and theme settings in one file
