+++ src/utils/Mantine.jsx
@@ -1,3 +1,31 @@
+import { Global } from "@mantine/core";
+
+const montserratUrl = "/fonts/Montserrat.ttf";
+const openSansUrl = "/fonts/OpenSans.ttf";
+
+const CustomFonts = () => {
+    return (
+        <Global styles={[
+            {
+                '@font-face': {
+                    fontFamily: 'Montserrat',
+                    src: `url('${montserratUrl}') format('truetype')`,
+                    fontStyle: 'normal',
+                    fontWeight: '200 700',
+                },
+            },
+            {
+                '@font-face': {
+                    fontFamily: 'Open Sans',
+                    src: `url('${openSansUrl}') format('truetype')`,
+                    fontStyle: 'normal',
+                    fontWeight: '200 700',
+                },
+            }
+        ]}/>
+    )
+}
+
 const mantineTheme = {
     colorScheme: 'dark',
     defaultRadius: 8,
@@ -45,4 +73,7 @@ const mantineTheme = {
     respectReducedMotion: true,
 }
 
-export default mantineTheme;
+export  {
+    CustomFonts,
+    mantineTheme
+}
