--- src/utils/Mantine/index.jsx
@@ -1,7 +1,0 @@
-import CustomFonts from "./MantineFont";
-import mantineTheme from "./MantineTheme";
-
-export  {
-    CustomFonts,
-    mantineTheme
-}
