+++ src/utils/Router.jsx
@@ -2,7 +2,7 @@ import { MantineProvider } from '@mantine/core';
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Landing } from '../pages';
-import { mantineTheme, CustomFonts } from "./Mantine";
+import { CustomFonts, mantineTheme } from './Mantine';
 
 const Router = () => {
      return (
